Rewrite content to use <img> instead of adding enclosures.
Adding images via enclosures doesn't appear to work in
Feedly, at least when using JSON Feed. I'm not sure if
that's because I shouldn't be using enclosures or because
the feeds package is mapping Item.Enclosure to the item's
"image" property instead of using the "attachment" property
as recommended at https://jsonfeed.org/mappingrssandatom.