Try to improve logic for including embedded tweets.
When handling replies, we used to include the first tweet
from the embedded tweet's /status page. This page sometimes
includes additional tweets: the tweet that the embedded
tweet is replying to and replies to the embedded tweet, I
think.

Change the logic so we look for a 'data-id' attribute
matching the ID from the URL (e.g. '123' in
twitter.com/status/123).